refactor #152: move legacy awards calculation to method
