+++ pages/api/scoreboard/index.ts
@@ -4,9 +4,12 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import { formatRoundsData } from '@/utils/formatRoundsData'
 import { calculateTotals } from '@/utils/calculateTotals'
 import { findTopFive } from '@/utils/findTopFive'
-import { mockApiData, tournamentAwards } from '@/constants'
+import { mockApiData } from '@/constants'
 import { calculateWinRate } from '@/utils/calculateWinRate'
-import { calculateCompletedChallenges } from '@/utils/calculateCompletedChallenges'
+import {
+  calculateCompletedChallenges,
+  calculateCompletedChallengesLegacy,
+} from '@/utils/calculateCompletedChallenges'
 import { removeBlacklistedPlayers } from '@/utils/removeBlacklistedPlayers'
 
 export default async function handler(_: NextApiRequest, res: NextApiResponse) {
@@ -49,27 +52,8 @@ export default async function handler(_: NextApiRequest, res: NextApiResponse) {
         playerStats.completedChallenges =
           calculateCompletedChallenges(GameAwardsSorted)
       } else {
-        // For retrocompatibility with data shape prev to 1.0.3 version
-        const awardValues = Object.entries(GameAwards).filter(
-          ([, value]) => value !== -1,
-        )
-        const generalValues = awardValues.filter(
-          ([key]) => !tournamentAwards.includes(key),
-        )
-        const tournamentValues = awardValues.filter(([key]) =>
-          tournamentAwards.includes(key),
-        )
-
-        playerStats.completedChallenges = {
-          general: [
-            generalValues.filter(([, value]) => value).length,
-            generalValues.length,
-          ],
-          tournament: [
-            tournamentValues.filter(([, value]) => value).length,
-            tournamentValues.length,
-          ],
-        }
+        playerStats.completedChallenges =
+          calculateCompletedChallengesLegacy(GameAwards)
       }
 
       playerStats.saves = GameStats.Saves
