+++ utils/calculateCompletedChallenges/index.ts
@@ -1,7 +1,15 @@
+import { tournamentAwards } from '@/constants'
 import { Challenges } from '@/interfaces/player'
 
 type RewardsData = Record<string, Record<string, number>>
 
+/**
+ * Calculates the awards a players has earned and return an object with two topuples
+ * cotaining the awards won and the total awards for tournamnts and standard games
+ *
+ * @param data player rewards
+ * @returns formatted player rewards
+ */
 export const calculateCompletedChallenges = (data: RewardsData): Challenges => {
   let trueRewards = 0
   let totalRewards = 0
@@ -32,5 +40,34 @@ export const calculateCompletedChallenges = (data: RewardsData): Challenges => {
   return {
     general: [trueRewards, totalRewards],
     tournament: [tournamentRewards[0], tournamentRewards[1]],
+  }
+}
+
+/**
+ * Calculates the awards a players has earned and return an object with two topuples
+ * cotaining the awards won and the total awards for tournamnts and standard games
+ * This method is valid for stats uploaded before 1.0.4 version
+ *
+ * @param data player rewards
+ * @returns formatted player rewards
+ */
+export const calculateCompletedChallengesLegacy = (data: []): Challenges => {
+  const awardValues = Object.entries(data).filter(([, value]) => value !== -1)
+  const generalValues = awardValues.filter(
+    ([key]) => !tournamentAwards.includes(key),
+  )
+  const tournamentValues = awardValues.filter(([key]) =>
+    tournamentAwards.includes(key),
+  )
+
+  return {
+    general: [
+      generalValues.filter(([, value]) => value).length,
+      generalValues.length,
+    ],
+    tournament: [
+      tournamentValues.filter(([, value]) => value).length,
+      tournamentValues.length,
+    ],
   }
 }
